+++ src/common/framework/util/collect_garbage.ts
