+++ src/webgpu/api/operation/resource_init/texture_zero_init_test.ts
@@ -1,6 +1,6 @@
 import * as C from '../../../../common/constants.js';
 import { TestCaseRecorder } from '../../../../common/framework/logging/test_case_recorder.js';
-import { params, poptions, pbool } from '../../../../common/framework/params.js';
+import { params, poptions, pbool } from '../../../../common/framework/params_builder.js';
 import { CaseParams } from '../../../../common/framework/params_utils.js';
 import { assert, unreachable } from '../../../../common/framework/util/util.js';
 import { kTextureAspects, kTextureFormatInfo, kTextureFormats } from '../../../capability_info.js';
