+++ src/webgpu/api/validation/createPipelineLayout.spec.ts
@@ -3,7 +3,7 @@ createPipelineLayout validation tests.
 `;
 
 import * as C from '../../../common/constants.js';
-import { pbool, poptions, params } from '../../../common/framework/params.js';
+import { pbool, poptions, params } from '../../../common/framework/params_builder.js';
 import { TestGroup } from '../../../common/framework/test_group.js';
 import {
   kBindingTypeInfo,
