+++ src/webgpu/api/validation/setBindGroup.spec.ts
@@ -2,7 +2,7 @@ export const description = `
 setBindGroup validation tests.
 `;
 
-import { poptions, params } from '../../../common/framework/params.js';
+import { poptions, params } from '../../../common/framework/params_builder.js';
 import { TestGroup } from '../../../common/framework/test_group.js';
 
 import { ValidationTest } from './validation_test.js';
